common/ns: Moved function c_run_set_namespaces from c_run to common/ns
The c_run_set_namespaces() function is moved to commen/ns.c.
This allows to use it from other places in cmld for future use cases,
which will be introduced in following patches.
Renamed function to more generic name 'ns_join_all()'.
